Oreshko_Vladislav_task1_load_json_and_xml/requirements.txt
apache-airflow==3.1.0
psycopg2==2.9.3
pandas==1.3.4
xmltodict==0.12.0